refactor common newsletter components into own file
